added a default config file to be used by client programs for default ros configuration. now using importlib instead of custom import_string() separated packagebound, confighandler, and config import classes simplified setup fixed tests
