+++ instance/pyros_setup.cfg
@@ -1,0 +1,11 @@
+# default configuration used when pyros-setup source is part of your source workspace, and you use it like :
+#
+# import pyros_setup
+# pyros_setup.configurable_import().configure().activate()
+#
+
+import os
+
+WORKSPACES=[os.path.join(os.path.dirname(__file__), '..', '..', 'devel')]
+# not specifying this means we use internal default
+DISTRO='indigo'
