+++ pyros_setup/config/__init__.py
@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .helpers import PackageBoundConfigHandler
+from .packagebound import PackageBound
+from .confighandler import ConfigHandler
+from .configimport import ConfigImport
 
 __all__ = [
-    'PackageBoundConfigHandler',
+    'PackageBound',
+    'ConfigHandler',
+    'ConfigImport',
 ]
